[REFACT] Add util packge for generic functions
